tests: rename PHP `Unit` to `Integration`
