+++ tests/php/Integration/MainTest.php
@@ -2,12 +2,12 @@
 /**
  * Main unit tests.
  *
- * @package OneSearch\Tests\Unit
+ * @package OneSearch\Tests\Integration
  */
 
 declare(strict_types = 1);
 
-namespace OneSearch\Tests\Unit;
+namespace OneSearch\Tests\Integration;
 
 use OneSearch\Main;
 use OneSearch\Tests\TestCase;
