+++ tests/php/Integration/Modules/Core/AssetsTest.php
@@ -2,12 +2,12 @@
 /**
  * Asset registration unit tests.
  *
- * @package OneSearch\Tests\Unit\Modules\Core
+ * @package OneSearch\Tests\Integration\Modules\Core
  */
 
 declare( strict_types = 1 );
 
-namespace OneSearch\Tests\Unit\Modules\Core;
+namespace OneSearch\Tests\Integration\Modules\Core;
 
 use OneSearch\Modules\Core\Assets;
 use OneSearch\Modules\Settings\Settings;
