+++ tests/php/Integration/Modules/Search/SearchTest.php
@@ -2,12 +2,12 @@
 /**
  * Search integration unit tests.
  *
- * @package OneSearch\Tests\Unit\Modules\Search
+ * @package OneSearch\Tests\Integration\Modules\Search
  */
 
 declare(strict_types = 1);
 
-namespace OneSearch\Tests\Unit\Modules\Search;
+namespace OneSearch\Tests\Integration\Modules\Search;
 
 use OneSearch\Modules\Search\Search;
 use OneSearch\Modules\Search\Settings as Search_Settings;
