+++ tests/php/Integration/UtilsTest.php
@@ -2,12 +2,12 @@
 /**
  * Utils unit tests.
  *
- * @package OneSearch\Tests\Unit
+ * @package OneSearch\Tests\Integration
  */
 
 declare( strict_types = 1 );
 
-namespace OneSearch\Tests\Unit;
+namespace OneSearch\Tests\Integration;
 
 use OneSearch\Tests\TestCase;
 use OneSearch\Utils;
